Move makeUniV2EdgeSwapQuote
